// frontend/src/features/flights/types.ts
export type FlightSearchParams = {
  q?: string;
  departure_id?: string;
  arrival_id?: string;
  gl?: string;
  hl?: string;
  currency?: string;
  type?: string;
  outbound_date?: string;
  return_date?: string;
  travel_class?: string;
  multi_city_json?: string;
  show_hidden?: string;
  exclude_basic?: string;
  deep_search?: string;
  adults?: string;
  children?: string;
  infants_in_seat?: string;
  infants_on_lap?: string;
  sort_by?: string;
  stops?: string;
  exclude_airlines?: string;
  include_airlines?: string;
  bags?: string;
  max_price?: string;
  outbound_times?: string;
  return_times?: string;
  emissions?: string;
  layover_duration?: string;
  exclude_conns?: string;
  max_duration?: string;
};